requirements.txt
Flask
Flask-Login
Flask-Bcrypt
Flask-WTF
FLask-Migrate
Flask-SQLAlchemy
Flask-Testing
python-decouple
python-dotenv
WTForms 
flask-wtf